feat(auth): added reset password page. Closes #671
