+++ libs/common/src/lib/constants.ts
@@ -23,6 +23,7 @@ export const PASSWORD_CONFIRM_URL = 'password.confirm';
 // URL for sending password forgotten requests
 export const EMAIL_PASSWORD_URL = 'password.email';
 export const PASSWORD_CONFIRMATION_URL = 'password.confirmation';
+export const PASSWORD_UPDATE_URL = 'password.update';
 export const FORGOT_PASSWORD_URL = 'password.request';
 export const TERMS_OF_SERVICE_URL = 'terms.show';
 export const PRIVACY_POLICY_URL = 'policy.show';
